+++ frontend/src/components/common/Circle/Circle.tsx
@@ -4,7 +4,7 @@ import { ReactComponent as CircleFill } from '@/assets/shape/circle-fill.svg';
 
 interface Props extends SVGProps<SVGSVGElement> {
   type: 'fill' | 'border';
-  color: string;
+  color?: string;
   size?: number;
 }
